Assignment/requirements.txt
yfinance
matplotlib.pyplot
matplotlib.animation